+++ feed.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="css/style.css" />
-    <script src="js/post/feed.js" type="module"></script>
+    <script src="js/socialFeed.js" type="module"></script>
     <script type="module" src="js/menu.js"></script>
     <title>feed | CSS-framework</title>
   </head>
@@ -191,6 +191,7 @@
               id="sortSelect"
               class="border rounded-md px-3 py-1.5 text-gray-700 focus:outline-none focus:ring-2 focus:ring-blue-500"
             >
+            <option value="[]">Select tag</option>
               <option value="travel">Travel</option>
               <option value="dog">dog</option>
               <option value="frontend">frontend</option>
